requirements.txt
Flask
Flask-DotEnv
requests
beautifulsoup4==4.12.2
python-steam-api==2.2.1